moving print fns to frontend.py
